requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==25.1.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.55.6
fsspec==2024.12.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.12.1
idna==3.10
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.8
libclang==18.1.1
lz4==4.4.3
MarkupSafe==2.1.5
matplotlib==3.10.0
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
networkx==3.2.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
pillow==11.1.0
psutil==6.1.1
py-cpuinfo==9.0.0
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
soupsieve==2.6
sympy==1.13.1
tensorflow_intel==2.18.0
termcolor==2.5.0
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tqdm==4.67.1
typing_extensions==4.9.0
tzdata==2025.1
ultralytics==8.3.68
ultralytics-thop==2.0.14
urllib3==2.3.0
wrapt==1.17.2
